requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.5
kombu==4.1.0
psycopg2==2.7.3
PyJWT==1.5.2
pymongo==3.5.1
pytz==2017.2
requests==2.18.3
urllib3==1.22
vine==1.1.4
